+++ example/integration_test/src/tests/unit/drawer_dialog_tests.dart
@@ -4,16 +4,10 @@ const _location = '========================== drawer_dialog_tests.dart';
 
 //
 Future<void> dialogTests(WidgetTester tester) async {
-  // Programmatically open the Drawer
-  ScaffoldState state = tester.firstState(find.byType(Scaffold));
+  //
+  final opened = await openDrawer(tester);
 
-  expect(state, isNotNull, reason: _location);
-
-  state.openDrawer();
-
-  await tester.pumpAndSettle();
-
-  if (state.isDrawerOpen) {
+  if (opened){
     //
     await _openDialogWindow('Show Box', tester);
 
@@ -21,7 +15,7 @@ Future<void> dialogTests(WidgetTester tester) async {
 
     await _openDialogWindow('Message Box', tester);
 
-    Navigator.pop(state.context);
+    Navigator.pop(App.context!);
 
     await tester.pumpAndSettle();
   }